Add filesystem web resource root contributors
Add ability to contribute filesystem web resource roots without
subclassing FilesystemWebResourceResolver. Move previous default
behavior in a contributor.

325310
