+++ org.eclipse.scout.rt.shared/src/main/java/org/eclipse/scout/rt/shared/ui/webresource/DefaultFilesystemWebResourceRootContributor.java
@@ -1,0 +1,49 @@
+/*
+ * Copyright (c) 2010-2022 BSI Business Systems Integration AG.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * https://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *     BSI Business Systems Integration AG - initial API and implementation
+ */
+package org.eclipse.scout.rt.shared.ui.webresource;
+
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.Collections;
+import java.util.List;
+
+public class DefaultFilesystemWebResourceRootContributor implements IFilesystemWebResourceRootContributor {
+  @Override
+  public List<Path> getRoots() {
+    return Collections.singletonList(findModuleRoot().resolve(AbstractWebResourceResolver.OUTPUT_FOLDER_NAME));
+  }
+
+  @SuppressWarnings("findbugs:NP_NULL_ON_SOME_PATH_FROM_RETURN_VALUE")
+  protected static Path findModuleRoot() {
+    Path workingDir = Paths.get("").toAbsolutePath();
+    Path parentDir = workingDir.getParent();
+    String folderName = workingDir.getFileName().toString();
+    String appModuleName = folderName;
+    if (folderName.endsWith(".dev") || folderName.endsWith("-dev")) {
+      appModuleName = folderName.substring(0, folderName.length() - 4);
+    }
+    Path resourceRoot = parentDir.resolve(appModuleName);
+    if (Files.isDirectory(resourceRoot) && Files.isReadable(resourceRoot)) {
+      return resourceRoot;
+    }
+
+    if (appModuleName.endsWith(".app") || appModuleName.endsWith("-app")) {
+      appModuleName = appModuleName.substring(0, appModuleName.length() - 4);
+    }
+    resourceRoot = parentDir.resolve(appModuleName);
+    if (Files.isDirectory(resourceRoot) && Files.isReadable(resourceRoot)) {
+      return resourceRoot;
+    }
+
+    return workingDir;
+  }
+}
